Added Filter and Import
Added filter to remove users without the EmailAddress field (Zendesk cannot create users without an email)
Importing the AWS Tools to Powershell.  Using an absolute path to support lower versions of PS.   PS 4.0 and above should be able to use "Import-Module AWSPowerShell".

In PowerShell 4.0 and later releases, Import-Module also searches the Program Files folder for installed modules, so it is not necessary to provide the full path to the module. You can run the following command to import the AWSPowerShell module. In PowerShell 3.0 and later, running a cmdlet in the module also automatically imports a module into your session.